Update Stone Tool Crafting to use #minecraft:st...
Updated Stone Tool Crafting to use #minecraft:stone_tool_materials
Rather than including modified recipes for stone tools and using a
custom tag from The 'strawberry' namespace, the Strawberry Pack now
modifies the #minecraft_stone_tool_materials item tag.

The tag was added in a snapshot at some point and this enables us to
do what previously requied five files in one.

In additon, Polished Stone Variants may no longer be used to be more
inline with the fact that Polished Blackstone is not a part of the
vanilla tag.