Added Lightning Griefing & Structures
New Gamerules which are on by default cause lightning to
behave in special ways

Lightnig Griefing
convert grass blocks to dirt and spawns Lightning Structures if
Lightnig Structures are enabled.

Lightnig Structures
When lighting strikes sand, it causes a spire of Fulgurite
(Stained Glass) to generate. Only if Lightning Griefing is ALSO
enabled though.

This may be expanded upon in the future.